# packages/contracts/openapi/k-admission.v1.yaml
openapi: 3.1.0
info:
  title: K-Admission University Data Plane API
  version: 1.4.0
  description: >
    Reference contract for a university-local K-Admission data plane.
    The university DB is the source of truth; central sync is not in the applicant critical path.


    1.2.0 aligns the contract with the repository implementation (spec discrepancy register
    D-7, D-16, D-17, D-19, D-20, D-22, D-24, D-26, D-32, D-34, D-35, D-37, D-38, D-39, D-40).
    Paths marked with `x-kadmission-register` were added by the implementation first.


    1.3.0 adds the common `429` response (`RateLimited`, `Retry-After`) to applicant operations (D-51).


    1.4.0 connects what the implementation had but the contract did not describe (D-55 to D-61):
    the application state machine is driven by validate / payment / finalize (READY, PAYMENT_PENDING,
    PAID are now real states and editing stops at PAYMENT_PENDING); one live payment per application
    (`200` resumes the open checkout, `409 PAYMENT_IN_PROGRESS` refuses a second payment); the form
    schema carries `profileFields` and `documents` so the client draws everything from config; the
    receipt is audited; server clock offset against the DB is reported (§A9); the applicant-facing
    common profile API on the central host; the active config includes its body and config drafts
    are linted (§A5); universities send the §04 heartbeat and the dashboard shows which university
...


    Common rules:
...
servers:
...
    variables:
...
paths:
...
  /internal/v1/documents/pending-scan:
    get:
      tags: [Internal]
      operationId: listDocumentsPendingScan
      x-kadmission-register: D-20
      summary: QUARANTINED documents waiting for the malware scanner
      description: Polled by document-service. The worker never writes the DB directly (ADR-0004).
      security: [{ mutualTLS: [] }]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: Pending documents
          content:
            application/json:
              schema:
                type: object
                required: [documents]
                properties:
                  documents:
                    type: array
                    items:
                      type: object
                      required: [documentId, objectKey, mediaType, sizeBytes, sha256]
                      properties:
                        documentId: { type: string, format: uuid }
                        objectKey: { type: string }
                        mediaType: { type: string }
                        sizeBytes: { type: integer, minimum: 0 }
                        sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
                        downloadUrl:
                          type: string
                          format: uri
                          description: >
                            Short-lived presigned GET for this object only (D-58). The worker holds
                            no object-storage credentials; the scan engine streams the file from here.
  /internal/v1/documents/{documentId}/scan-result:
    post:
      tags: [Internal]
      operationId: reportDocumentScanResult
      x-kadmission-register: D-20
      security: [{ mutualTLS: [] }]
      parameters:
        - $ref: '#/components/parameters/DocumentId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [result]
              properties:
                result: { type: string, enum: [CLEAN, MALICIOUS, ERROR] }
                scanner: { type: string, description: Engine name recorded in the evidence (D-58) }
                engineVersion: { type: string, description: Engine and signature database version }
                signature: { type: string, maxLength: 200, description: What the engine found (MALICIOUS only) }
      responses:
        '200':
          description: Scan result applied
          content:
            application/json:
              schema:
                type: object
                required: [documentId, result, status]
                properties:
                  documentId: { type: string, format: uuid }
                  result: { type: string, enum: [CLEAN, MALICIOUS, ERROR] }
                  status: { type: string, enum: [UPLOADING, QUARANTINED, AVAILABLE, REJECTED, DELETED] }
        '400':
          $ref: '#/components/responses/Problem'
...
  /api/v1/profile:
    get:
      tags: [Central]
      operationId: getMyProfile
      x-kadmission-register: D-57
      summary: My common profile and per-university release consents
      description: >
        The common profile is written once and copied, with consent, into each university
        application as a point-in-time snapshot (§10 §3). Identity comes only from the
        authenticated subject (`x-authenticated-subject` behind the gateway; `x-subject-token`
        in dev-headers mode, refused in production). No subject token in the body or URL.
      servers:
        - url: https://central.k-admission.kr
      parameters:
        - $ref: '#/components/parameters/SubjectToken'
      responses:
        '200':
          description: Profile (empty fields when none was saved)
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommonProfile' }
        '400':
          $ref: '#/components/responses/Problem'
    put:
      tags: [Central]
      operationId: replaceMyProfile
      x-kadmission-register: D-57
      summary: Replace my common profile and release consents
      description: >
        Full replacement — sending the same body again gives the same result, so no
        Idempotency-Key is needed. Universities missing from `consents` have their consent
        revoked (recorded, not deleted); snapshots already copied into applications do not
        change. Consent can only name saved fields. Only the platform-standard common profile
        fields are accepted (highSchool, graduationYear, contactEmail, phone — contracts
        COMMON_PROFILE_FIELDS); an empty string or null removes a field.
      servers:
        - url: https://central.k-admission.kr
      parameters:
        - $ref: '#/components/parameters/SubjectToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fields]
              properties:
                fields:
                  type: object
                  additionalProperties: false
                  properties:
                    highSchool: { type: [string, 'null'], maxLength: 100 }
                    graduationYear: { type: [integer, 'null'], minimum: 1990, maximum: 2100 }
                    contactEmail: { type: [string, 'null'], maxLength: 254 }
                    phone: { type: [string, 'null'], maxLength: 20 }
                consents:
                  type: array
                  items:
                    type: object
                    required: [universityId, fieldCodes]
                    properties:
                      universityId: { type: string }
                      fieldCodes:
                        type: array
                        items: { type: string }
      responses:
        '200':
          description: Saved profile
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommonProfile' }
        '400':
          $ref: '#/components/responses/Problem'
  /internal/v1/profile-snapshots:
    post:
      tags: [Central]
      operationId: releaseProfileSnapshot
      x-kadmission-register: D-17
      summary: Common Profile Vault — release consented fields to a university
      description: >
        Called by a university data plane when creating an application. Only the intersection of
        requested and consented fields leaves the vault; requested-but-unconsented fields are
        named in `withheldFields` instead of silently dropped. Best-effort: if central does not
        answer, the application is created empty (D-18). `applicationRef` is an opaque
        university-salted value, never the university's internal applicant id (D-39).
      servers:
        - url: https://central.k-admission.kr
      security: [{ mutualTLS: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subjectToken, universityId, requestedFields]
              properties:
                subjectToken: { type: string }
                universityId: { type: string }
                requestedFields:
                  type: array
                  items: { type: string }
                applicationRef: { type: string }
      responses:
        '200':
          description: Snapshot
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema:
                type: object
                required: [fields, releasedFields, withheldFields, releasedAt]
                properties:
                  fields: { type: object, additionalProperties: true }
                  releasedFields:
                    type: array
                    items: { type: string }
                  withheldFields:
                    type: array
                    items: { type: string }
                  releasedAt: { type: string, format: date-time }
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://id.example.kr/.well-known/openid-configuration
    mutualTLS:
      type: mutualTLS
  parameters:
...
    DocumentId:
      in: path
      name: documentId
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 16, maxLength: 200 }
...
    SubjectToken:
      in: header
      name: x-subject-token
      required: true
      description: Applicant subject token. Header only — never in the URL (D-39).
      schema: { type: string, minLength: 1 }
  responses:
    Problem:
      description: Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    CommonProfile:
      type: object
      required: [fields, consents, updatedAt]
      properties:
        fields: { type: object, additionalProperties: true }
        consents:
          type: array
          items:
            type: object
            required: [universityId, fieldCodes, grantedAt]
            properties:
              universityId: { type: string }
              fieldCodes:
                type: array
                items: { type: string }
              grantedAt: { type: string, format: date-time }
        updatedAt: { type: [string, 'null'], format: date-time }
...
    Problem:
      type: object
      required: [type, title, status, code, traceId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        traceId: { type: string }
        serverTime: { type: string, format: date-time }
security:
  - oidc: []